Migrate icon-maker to appdir
